requirements.txt
langgraph>=0.2.34
langchain-core>=0.3.7
langchain-openai>=0.2.6
langchain-community>=0.3.3
tavily-python>=0.3.0
python-dotenv>=1.0.1
httpx>=0.27.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
lxml>=5.1.0

